+++ pkg/storage/storagewrappers/sharediterator/shared_iterator_datastore.go
@@ -2,7 +2,6 @@ package sharediterator
 
 import (
 	"context"
-	"fmt"
 	"strconv"
 	"sync"
 	"sync/atomic"
@@ -554,35 +553,32 @@ const bufferSize = 100
 // The singleflight.Group type was used as a comparison to the await type, but was found to be ~59% slower than await
 // in concurrent stress test benchmarks.
 type await struct {
-	executing atomic.Bool
-	ch        atomic.Pointer[chan struct{}]
+	active bool
+	wg     *sync.WaitGroup
+	mu     sync.Mutex
 }
 
 // Do executes the provided function fn if it is not already being executed.
-// The channel returned will be closed once the execution has completed.
-func (a *await) Do(fn func()) chan struct{} {
-	i := make(chan struct{})
-	a.ch.CompareAndSwap(nil, &i)
-	ch := *a.ch.Load()
-
-	if !a.executing.Swap(true) {
-		// If we are not already executing, we start a new goroutine to execute.
-		go func() {
-			fn()
-
-			// Important! The executing state must be set to false before closing the old channel.
-			// This avoids a race condition in which a goroutine waiting on the channel close might
-			// enter this function again and try to execute before the execution state is reset.
-			a.executing.Store(false)
-
-			newCh := make(chan struct{})
-			currentCh := a.ch.Swap(&newCh)
-
-			// Close the old channel to signal waiting goroutines to wake up.
-			close(*currentCh)
-		}()
+// The first goroutine to call Do will execute the function, while subsequent calls will block until the function has completed.
+// This ensures that only one goroutine can execute the function at a time, preventing concurrent execution of the function.
+func (a *await) Do(fn func()) {
+	a.mu.Lock()
+	if a.active {
+		a.mu.Unlock()
+		a.wg.Wait()
+		return
 	}
-	return ch
+	a.active = true
+	a.wg = new(sync.WaitGroup)
+	a.wg.Add(1)
+	a.mu.Unlock()
+
+	fn()
+	a.wg.Done()
+
+	a.mu.Lock()
+	a.active = false
+	a.mu.Unlock()
 }
 
 // iteratorReader is a wrapper around a storage.Iterator that implements the reader interface.
@@ -594,20 +590,14 @@ type iteratorReader[T any] struct {
 // The method will read up to the length of the buffer, and if there are fewer items available,
 // it will return the number of items read and an error if any occurred.
 func (ir *iteratorReader[T]) Read(ctx context.Context, buf []T) (int, error) {
-	var i int
-	var e error
-
-	buflen := len(buf)
-
-	for ; i < buflen; i++ {
-		t, ierr := ir.Next(ctx)
-		if ierr != nil {
-			e = ierr
-			break
+	for i := range buf {
+		t, err := ir.Next(ctx)
+		if err != nil {
+			return i, err
 		}
 		buf[i] = t
 	}
-	return i, e
+	return len(buf), nil
 }
 
 // iteratorState holds the state of the shared iterator.
@@ -664,26 +654,9 @@ func newSharedIterator(it storage.TupleIterator, cleanup func()) *sharedIterator
 
 	var closed atomic.Bool
 
-	// Initialize the channel that will be used to signal when new items are available.
-	ch := make(chan struct{})
-	var pch atomic.Pointer[chan struct{}]
-	pch.Store(&ch)
-
 	// Initialize the reference counter to 1, indicating that there is one active instance of the iterator.
 	var refs atomic.Int64
 	refs.Store(1)
-
-	// Initialize the error pointer to nil, indicating that there are no errors at the start.
-	// This will be used to store any errors that occur during the iteration.
-	var err atomic.Pointer[error]
-	terr := new(error)
-	err.Store(terr)
-
-	// Initialize the items pointer to an empty slice of tuples.
-	// This will be used to store the items fetched by the iterator.
-	var items atomic.Pointer[[]*openfgav1.Tuple]
-	titems := make([]*openfgav1.Tuple, 0)
-	items.Store(&titems)
 
 	ir := &iteratorReader[*openfgav1.Tuple]{
 		Iterator: it,
@@ -694,7 +667,7 @@ func newSharedIterator(it storage.TupleIterator, cleanup func()) *sharedIterator
 	var pstate atomic.Pointer[iteratorState]
 	pstate.Store(&state)
 
-	newIter := sharedIterator{
+	return &sharedIterator{
 		await:   &aw,
 		cleanup: cleanup,
 		ir:      ir,
@@ -702,8 +675,6 @@ func newSharedIterator(it storage.TupleIterator, cleanup func()) *sharedIterator
 		refs:    &refs,
 		closed:  &closed,
 	}
-
-	return &newIter
 }
 
 // clone creates a new shared iterator that shares the same context, cancellation function, and other fields.
@@ -731,56 +702,46 @@ func (s *sharedIterator) clone() *sharedIterator {
 	}
 }
 
+// fetchMore is a method that fetches more items from the underlying storage.TupleIterator.
+// It reads a fixed number of items (bufferSize) from the iterator and appends them
+// to the shared items slice in the iterator state.
+// If an error occurs during the read operation, it updates the error in the iterator state.
+func (s *sharedIterator) fetchMore() {
+	var buf [bufferSize]*openfgav1.Tuple
+	read, e := s.ir.Read(context.Background(), buf[:])
+
+	// Load the current items from the shared items pointer and append the newly fetched items to it.
+	state := s.state.Load()
+
+	newState := &iteratorState{
+		items: make([]*openfgav1.Tuple, len(state.items)+read),
+		err:   state.err,
+	}
+
+	copy(newState.items, state.items)
+	copy(newState.items[len(state.items):], buf[:read])
+
+	if e != nil {
+		newState.err = e
+	}
+
+	s.state.Store(newState)
+}
+
 // fetchAndWait is a method that fetches items from the underlying storage.TupleIterator and waits for new items to be available.
 // It blocks until new items are fetched or an error occurs.
 // The items and err pointers are updated with the fetched items and any error encountered.
-func (s *sharedIterator) fetchAndWait(ctx context.Context, items *[]*openfgav1.Tuple, err *error) {
-	// Iterate until we have items available or an error occurs.
+func (s *sharedIterator) fetchAndWait(items *[]*openfgav1.Tuple, err *error) {
 	for {
-		if ctx.Err() != nil {
+		state := s.state.Load()
+
+		if s.head < len(state.items) || state.err != nil {
+			*items = state.items
+			*err = state.err
 			return
 		}
 
-		state := *s.state.Load()
-		*items = state.items
-		*err = state.err
-
-		if s.head < len(*items) || *err != nil {
-			break
-		}
-
-		ch := s.await.Do(func() {
-			defer func() {
-				if r := recover(); r != nil {
-					var err error
-					if e, ok := r.(error); ok {
-						err = fmt.Errorf("recovered from panic: %w", e)
-					} else {
-						err = fmt.Errorf("recovered from panic: %v", r)
-					}
-					state := *s.state.Load()
-					state.err = err
-					s.state.Store(&state)
-				}
-			}()
-
-			var buf [bufferSize]*openfgav1.Tuple
-			read, e := s.ir.Read(context.Background(), buf[:])
-
-			// Load the current items from the shared items pointer and append the newly fetched items to it.
-			state := *s.state.Load()
-			state.items = append(state.items, buf[:read]...)
-
-			if e != nil {
-				state.err = e
-			}
-			s.state.Store(&state)
-		})
-
-		select {
-		case <-ctx.Done():
-		case <-ch:
-		}
+		s.await.Do(s.fetchMore)
 	}
 }
 
@@ -800,7 +761,7 @@ func (s *sharedIterator) current(ctx context.Context) (*openfgav1.Tuple, error) 
 	var items []*openfgav1.Tuple
 	var err error
 
-	s.fetchAndWait(ctx, &items, &err)
+	s.fetchAndWait(&items, &err)
 
 	if ctx.Err() != nil {
 		return nil, ctx.Err()